Revert "Added changes for making tree view persistent, made changes for bugs for loading screen (#153) (#167)"
This reverts commit 2269d1de55873286a5d1821fc0f220cd967a7aff.
