+++ common/constants/index.ts
@@ -10,14 +10,8 @@ export const OPENSEARCH_ACC_DOCUMENTATION_URL =
 export const ACC_INDEX_TYPE_DOCUMENTATION_URL =
   'https://github.com/opensearch-project/opensearch-spark/blob/main/docs/index.md';
 
-export const TREE_ITEM_SKIPPING_INDEX_DEFAULT_NAME = `skipping_index`;
-export const TREE_ITEM_COVERING_INDEX_DEFAULT_NAME = `covering_index`;
-export const TREE_ITEM_MATERIALIZED_VIEW_DEFAULT_NAME = `materialized_view`;
-export const TREE_ITEM_DATABASE_NAME_DEFAULT_NAME = `database`;
-export const TREE_ITEM_TABLE_NAME_DEFAULT_NAME = `table`;
-export const TREE_ITEM_LOAD_MATERIALIZED_BADGE_NAME = `Load Materialized View`;
-export const TREE_ITEM_BADGE_NAME =`badge`
-export const LOAD_OPENSEARCH_INDICES_QUERY = `SHOW tables LIKE '%';`;
+export const SKIPPING_INDEX = `skipping_index`;
+export const ON_LOAD_QUERY = `SHOW tables LIKE '%';`;
 export const SKIPPING_INDEX_QUERY = `CREATE SKIPPING INDEX ON myS3.logs_db.http_logs 
 (status VALUE_SET) 
 WITH (
@@ -90,5 +84,3 @@ export const ACCELERATION_INDEX_NAME_INFO = `All OpenSearch acceleration indices
   `;
 
 export const SIDEBAR_POLL_INTERVAL_MS = 5000;
-
-export const FETCH_OPENSEARCH_INDICES_PATH = '/api/sql_console/sqlquery'
